backend/pyproject.toml
[tool.poetry]
name = "test-case"
version = "0.1.0"
description = ""
authors = ["vellimir <kkazancov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
